refactor(redeem-ops): split entitlementService by responsibility (P3-2)
entitlementService was a 1,417-line factory holding four unrelated jobs. Each
now lives in its own module, lifted verbatim:

  entitlementDelivery.js        email + WhatsApp fan-out and its receipts
  entitlementQuery.js           the masked list projection
  entitlementReconciliation.js  the four catch-up sweeps
  redemptionEvents.js           the redemption_events writer (shared)

entitlementService.js keeps the state machine — issue, unlock, undo, manual
issue, cancel, cascade-cancel — and drops 1,417 → 960 lines.

The public factory API is unchanged, key by key, and so is `flushDeliveries`:
the in-flight set moved into entitlementDelivery.js and is re-exported from its
old home, because tests and the sweeps use it as a barrier.

The sweeps take issueForProspect, queueDelivery and writeEvent as injected
functions rather than importing them. That is deliberate — a reconciliation
sweep should drive the live path, never grow its own copy of it, which is how
the trial/physical/draw fork ended up re-implemented across this file.

The DRY win the task flagged: entitlementService and redemptionService each
carried a private writeEvent, byte-identical apart from the default actorType
('system' for the engine, 'staff' for the counter). That default is the whole
meaning of an audit row, and two copies is exactly how it drifts. One function
now, with the default as a parameter. It takes `d` rather than the model so the
lookup stays late-bound — both services build the writer once but let tests swap
d.RedemptionEvent afterwards.

Proof is the existing suites passing with ZERO test edits: 53 redeem-ops /
entitlement / draw suites (766 tests), unit 129/129 (2,202 tests), and the full
integration tier 148/148 (2,600 tests).
